requirements.txt
websockets>=12.0
aiohttp>=3.9.0
pyaudio>=0.2.13
numpy>=1.24.0
python-dotenv>=1.0.0
click>=8.0.0
pydub>=0.25.1
openai>=1.0.0
nest-asyncio>=1.5.0